requirements.txt
tweepy
google-generativeai
python-dotenv
schedule
